requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
certifi @ file:///C:/ci/certifi_1655907258560/work/certifi
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.109.2
h11==0.14.0
idna==3.6
OpenOPC-DA==1.5.0
packaging==23.2
pefile==2023.2.7
pillow==10.2.0
pydantic==2.6.1
pydantic_core==2.16.2
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
pymongo==4.6.2
Pyro5==5.15
pywin32==306
pywin32-ctypes==0.2.2
schedule==1.2.1
serpent==1.41
sniffio==1.3.0
starlette==0.36.3
ttkbootstrap==1.10.1
typing_extensions==4.9.0
uvicorn==0.27.1
wincertstore==0.2
